fix: macro not updating
This fix changes the logic on how the macro is found and generated.
Previously, the code would try to find the macro based off a macro index
and try and do a string.find on RHT.b. Seeing RHT.b is a defunct part of
the macro, I changed this to use GetMacroIndexByName as the macro name
is always 'RHT'. This allows us to set the index in a much cleaner
fashion. Lastly, I removed the last line of SetRandomHearthToy that was
setting the RHT.b button which did not exist. This has been extensively
tested and the bug is now resolved